fix logic

+++ src/main/java/qnt/moviebooking/dto/request/VoucherRequestDto.java
@@ -1,5 +1,6 @@
 package qnt.moviebooking.dto.request;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import lombok.AllArgsConstructor;
@@ -13,9 +14,9 @@ import lombok.NoArgsConstructor;
 @Builder
 public class VoucherRequestDto {
     private String code;
-    private Double discountAmount;
+    private BigDecimal discountAmount;
     private Double discountPercentage;
-    private Double discountMax;
+    private BigDecimal discountMax;
     private LocalDateTime expiryDate;
     private boolean active;
 }
